added remove_duplicates_vas_function.R
R function that can be used to remove duplicate items. Future versions of vas.praat will have a column built in to allow for easy extraction of the items that were duplicated for reliability purposes. This function can be used in the meantime. It takes as input the R dataframe object corresponding to the .csv output produced by vas.praat, and returns two dataframes as output: output_unique, which is the original data with duplicates removed, and output_duplicate, which is a dataframe of the same structure containing just the duplicates. The "unique" items correspond to the first presentation to the listener, and the "duplicate" items correspond to the second presentation.